move transitions into animated calendar
